fix: sticky dialog footers on mobile + UI refinements
Pin save/cancel/delete buttons to bottom of dialog viewport on mobile so
users don't have to scroll to reach actions. Applies to accommodation,
transportation, activity, and reservation dialogs. Also fixes double-scroll
in ActivityDialog and broken overflow containment in ReservationDialog.

Includes miscellaneous sidebar, timeline, and panel refinements from the
newStyle branch.
