+++ src/components/trip/timeline/TripDatesPanel.tsx
@@ -3,17 +3,18 @@ import React from "react";
 import { Button } from "@/components/ui/button";
 import { Edit } from "lucide-react";
 import Header from "@/components/trip/_shared/Header";
-import { parse, differenceInCalendarDays, format } from "date-fns";
+import { differenceInCalendarDays, format } from "date-fns";
 
 interface TripDatesPanelProps {
   trip: { arrival_date: string; departure_date: string } | null;
   onEdit: () => void;
+  canEdit?: boolean;
   isMobile: boolean;
   onClose: () => void;
   onBack: () => void;
 }
 
-/** Parse an ISO “YYYY-MM-DD” string to a local‐midnight Date */
+/** Parse an ISO "YYYY-MM-DD" string to a local-midnight Date */
 function parseLocalDate(iso: string): Date {
   const [y, m, d] = iso.split("-").map(Number);
   return new Date(y, m - 1, d);
@@ -22,6 +23,7 @@ function parseLocalDate(iso: string): Date {
 export default function TripDatesPanel({
   trip,
   onEdit,
+  canEdit = true,
   isMobile,
   onClose,
   onBack,
@@ -36,14 +38,16 @@ export default function TripDatesPanel({
     <div className="p-4">
       <Header title="Trip Dates" isMobile={isMobile} onBack={onBack} onClose={onClose} />
 
-      <Button
-        size="sm"
-        onClick={onEdit}
-        className="mb-4 w-full bg-earth-500 text-white hover:bg-earth-600"
-      >
-        <Edit size={14} className="mr-1" />
-        Edit Dates
-      </Button>
+      {canEdit && (
+        <Button
+          size="sm"
+          onClick={onEdit}
+          className="mb-4 w-full bg-earth-500 text-white hover:bg-earth-600"
+        >
+          <Edit size={14} className="mr-1" />
+          Edit Dates
+        </Button>
+      )}
 
       <div className="space-y-3">
         <div className="rounded-lg bg-sand-50 p-3">
@@ -61,7 +65,7 @@ export default function TripDatesPanel({
         <div className="rounded-lg bg-sand-50 p-3">
           <p className="text-sm font-medium text-earth-600">Duration</p>
           <p className="text-sm text-sand-700">
-            {nights > 0 ? `${nights} night${nights === 1 ? "" : "s"}` : "—"}
+            {nights > 0 ? `${nights} night${nights === 1 ? "" : "s"}` : "\u2014"}
           </p>
         </div>
       </div>
